Refactoring test_if_hostname_exists_function
