+++ response_from_executed_request.txt
@@ -1,9 +1,1 @@
 
-HTTP/1.1 200 OK
-Connection: keep-alive
-Content-Length: 301
-Content-Type: text/plain
-Date: Wed, 25 Sep 2019 01:20:24 GMT
-Last-Modified: Tue, 20 Mar 2018 23:57:23 GMT
-Server: OpenBSD httpd
-
