+++ run.sh
@@ -10,6 +10,6 @@ echo "" > response_from_executed_request.txt
 
 #./exe GET 10.132.20.22 80 /aulas/gr/readme.txt response_from_executed_request.txt
 
-#./exe GET indoorweb.com.br 80 / response_from_executed_request.txt
-./exe HEAD 10.132.20.22 80 /aulas/gr/readme.txt response_from_executed_request.txt
+./exe GET indoorweb.com.br 80 / response_from_executed_request.txt
+#./exe HEAD 10.132.20.22 80 /aulas/gr/readme.txt response_from_executed_request.txt
